# backend/openapi.json
{
  "components": {
    "schemas": {
      "AvailabilityRead": {
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "party_size": {
            "title": "Party Size",
            "type": "integer"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableRead"
            },
            "title": "Tables",
            "type": "array"
          }
        },
        "required": [
          "start_at",
          "end_at",
          "party_size",
          "tables"
        ],
        "title": "AvailabilityRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_RestaurantRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RestaurantRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RestaurantRead]",
        "type": "object"
      },
...
      "RestaurantCreate": {
        "properties": {
          "city": {
            "maxLength": 80,
            "minLength": 1,
            "title": "City",
            "type": "string"
          },
          "closes_at": {
            "format": "time",
            "title": "Closes At",
            "type": "string"
          },
          "default_duration_minutes": {
            "default": 90,
            "maximum": 480.0,
            "minimum": 15.0,
            "title": "Default Duration Minutes",
            "type": "integer"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "opens_at": {
            "format": "time",
            "title": "Opens At",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "name",
          "city",
          "opens_at",
          "closes_at"
        ],
        "title": "RestaurantCreate",
        "type": "object"
      },
      "RestaurantRead": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "closes_at": {
            "format": "time",
            "title": "Closes At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_duration_minutes": {
            "title": "Default Duration Minutes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opens_at": {
            "format": "time",
            "title": "Opens At",
            "type": "string"
          },
          "table_count": {
            "default": 0,
            "title": "Table Count",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "city",
          "timezone",
          "opens_at",
          "closes_at",
          "default_duration_minutes",
          "created_at"
        ],
        "title": "RestaurantRead",
        "type": "object"
      },
      "RestaurantUpdate": {
        "description": "Partial update: only the fields that are sent change; `null` is not a valid value.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "closes_at": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closes At"
          },
          "default_duration_minutes": {
            "anyOf": [
              {
                "maximum": 480.0,
                "minimum": 15.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration Minutes"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "opens_at": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opens At"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "RestaurantUpdate",
        "type": "object"
      },
      "SlotRead": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "free_tables": {
            "title": "Free Tables",
            "type": "integer"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "start_at",
          "end_at",
          "local_time",
          "available",
          "free_tables"
        ],
        "title": "SlotRead",
        "type": "object"
      },
      "SlotsRead": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "party_size": {
            "title": "Party Size",
            "type": "integer"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotRead"
            },
            "title": "Slots",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "date",
          "timezone",
          "duration_minutes",
          "party_size",
          "slots"
        ],
        "title": "SlotsRead",
        "type": "object"
      },
      "TableCreate": {
        "properties": {
          "capacity": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Capacity",
            "type": "integer"
          },
          "label": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "capacity"
        ],
        "title": "TableCreate",
        "type": "object"
      },
      "TableRead": {
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "restaurant_id": {
            "title": "Restaurant Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "restaurant_id",
          "label",
          "capacity",
          "is_active"
        ],
        "title": "TableRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "TableFlow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/restaurants": {
      "get": {
        "operationId": "list_restaurants_api_v1_restaurants_get",
        "parameters": [
          {
            "description": "Exact city, case-insensitive",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact city, case-insensitive",
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RestaurantRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Restaurants",
        "tags": [
          "restaurants"
        ]
      },
      "post": {
        "operationId": "create_restaurant_api_v1_restaurants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestaurantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Restaurant",
        "tags": [
          "restaurants"
        ]
      }
    },
...
    "/api/v1/restaurants/{restaurant_id}": {
      "get": {
        "operationId": "get_restaurant_api_v1_restaurants__restaurant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Restaurant",
        "tags": [
          "restaurants"
        ]
      },
      "patch": {
        "operationId": "update_restaurant_api_v1_restaurants__restaurant_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestaurantUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Restaurant",
        "tags": [
          "restaurants"
        ]
      }
    },
    "/api/v1/restaurants/{restaurant_id}/availability": {
      "get": {
        "operationId": "get_availability_api_v1_restaurants__restaurant_id__availability_get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End At"
            }
          },
          {
            "in": "query",
            "name": "party_size",
            "required": true,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "title": "Party Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Availability",
        "tags": [
          "restaurants"
        ]
      }
    },
    "/api/v1/restaurants/{restaurant_id}/availability/slots": {
      "get": {
        "description": "Start times for a day (every 30 minutes) with how many tables are free for the party.",
        "operationId": "get_slots_api_v1_restaurants__restaurant_id__availability_slots_get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "party_size",
            "required": true,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "title": "Party Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Slots",
        "tags": [
          "restaurants"
        ]
      }
    },
    "/api/v1/restaurants/{restaurant_id}/tables": {
      "get": {
        "operationId": "list_tables_api_v1_restaurants__restaurant_id__tables_get",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TableRead"
                  },
                  "title": "Response List Tables Api V1 Restaurants  Restaurant Id  Tables Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tables",
        "tags": [
          "restaurants"
        ]
      },
      "post": {
        "operationId": "create_table_api_v1_restaurants__restaurant_id__tables_post",
        "parameters": [
          {
            "in": "path",
            "name": "restaurant_id",
            "required": true,
            "schema": {
              "title": "Restaurant Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Table",
        "tags": [
          "restaurants"
        ]
      }
    },
...
  }
}